options: make --loop always per-file
In this example, only f1.mkv was looped: mpv f1.mkv f2.mkv --loop=2
This is because the playloop actually changes the global option value,
assuming it would be reset when going to the next file. When mpv was
changed to not reset options between files, this assumption was broken.